restructured app into packages
