+++ .gitignore
@@ -2,6 +2,7 @@ backend/.env
 backend/path/
 backend/.venv/
 backend/__pycache__
+backend/app/__pycache__
 frontend/.env
 frontend/dist/
 frontend/node_modules/
