+++ backend/app/database/__init__.py
@@ -1,0 +1,3 @@
+from app.database.driver import Neo4jDriver
+from app.database.embedding import Embedding
+from app.database.chat_history import MessageHistory
