+++ backend/app/retrieval/__init__.py
@@ -1,0 +1,1 @@
+from app.retrieval.retriever import Retriever
